bug(filenames):implement-lfa-feedback rename all file to use camel case [Finishes #165555770]
