+++ server/controllers/accountsController.js
@@ -1,7 +1,7 @@
 /* eslint-disable consistent-return */
-import User from '../models/user.model';
-import Account from '../models/account.model';
-import Transaction from '../models/transaction.model';
+import User from '../models/user';
+import Account from '../models/account';
+import Transaction from '../models/transaction';
 import validateCreateBankAccountInput from '../validation/bankAccount';
 
 const AccountsController = {
@@ -132,14 +132,12 @@ const AccountsController = {
       return res.status(200).json({ status: 200, data: transactions.rows });
     } catch (err) {
       if (err) {
-        console.log(err)
         return res.status(404).json({
           status: 404,
           error: 'Acount does not exist',
         });
       }
     }
-    // return res.status(200).json({ status: 200, data: '' });
   },
 
   /**
