+++ server/controllers/usersController.js
@@ -1,9 +1,9 @@
 /* eslint-disable consistent-return */
 import jwt from 'jsonwebtoken';
 import { config } from 'dotenv';
-import User from '../models/user.model';
+import User from '../models/user';
 import { hash, setAuthToken } from '../utils/helpers';
-import validateSignUpInput from '../validation/authentication/signup';
+import validateSignUpInput from '../validation/authentication/signUp';
 
 config();
 
@@ -121,7 +121,7 @@ const UsersController = {
           }
           if (type === 'admin') {
             User.findOneAndUpdate('id', payload.id, { is_admin: true })
-            .then(() => {});
+              .then(() => {});
           }
 
           // eslint-disable-next-line consistent-return
