+++ server/database/seeders/index.js
@@ -7,15 +7,15 @@ const log = debug('express:migration');
 
 (function runSeeder() {
   setTimeout(() => {
-  	userSeeder();
-  	log('seeded users table successfully at', new Date());
+    userSeeder();
+    log('seeded users table successfully at', new Date());
   }, 100);
   setTimeout(() => {
-  	accountSeeder();
-  	log('seeded accounts table successfully at', new Date());
+    accountSeeder();
+    log('seeded accounts table successfully at', new Date());
   }, 2000);
   setTimeout(() => {
-  	transactionSeeder();
-  	log('seeded transactions table successfully at', new Date());
+    transactionSeeder();
+    log('seeded transactions table successfully at', new Date());
   }, 3000);
 }());
