+++ server/database/seeders/usersTableSeeder.js
@@ -1,15 +1,15 @@
 /* eslint-disable no-restricted-syntax */
 import { user } from '../factories/userFactory';
 import DB from '../../config/database';
-import User from '../../models/user.model';
+import User from '../../models/user';
 
-export default function userSeeder() {
+const userSeeder = () => {
   DB.query('TRUNCATE users CASCADE')
     .then(() => {
       for (const data of user) {
-        User.save(data).then(result => result.rows)
+        User.save(data).then(result => result.rows);
       }
     });
-}
+};
 
-userSeeder()
+export default userSeeder;
