--- server/db/users.db.js
@@ -1,17 +1,0 @@
-/*
-** Data structure
-** to mimick database
-*/
-
-const users = [
-  {
-    id: 1,
-    email: 'mark55@email.com',
-    firstName: 'Mark',
-    lastName: 'Henry',
-    password: 'markhen',
-  },
-
-];
-
-export default users;
