+++ server/routes/transactionsRoute.js
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import Auth from '../middleware/verifyToken';
-import TransactionsController from '../controllers/transactions.controller';
+import TransactionsController from '../controllers/transactionsController';
 
 const router = Router();
