+++ server/utils/helpers.js
@@ -18,13 +18,13 @@ export const setAuthToken = (req, token) => {
 };
 
 // Hash password
-export function hash(password) {
-  return new Promise((resolve, reject) => {
+export const hash = password => (
+  new Promise((resolve, reject) => {
     bcrypt.genSalt(10, (_err, salt) => {
       bcrypt.hash(password, salt, (err, hashed) => {
         if (err) reject(err);
         resolve(hashed);
       });
     });
-  });
-}
+  })
+);
