+++ server/validation/authentication/signUp.js
